# [MRG] Faster `before_run`