… brian2/spatialneuron/spatialneuron.py
… brian2/stateupdaters/base.py
… brian2/stateupdaters/explicit.py
… brian2/stateupdaters/exponential_euler.py
… brian2/synapses/synapses.py
… brian2/tests/test_namespaces.py
… brian2/tests/test_parsing.py
… brian2/tests/test_refractory.py
… brian2/tests/test_stateupdaters.py
… docs_sphinx/developer/equations_namespaces.rst