As we discussed, instead of implementing joblib caching (#124) I first had a look at where the preparation time is spend. In this branch, I implemented a few small uncontroversial changes (e.g. `check_code_units` was called before `create_runner_codeobj`, but the latter called `check_code_units` again), but also some bigger changes that are at least worth discussing:
1. Instead of passing down the level argument down from `before_run` all the way until when namespaces are actually resolved, we get the local namespace in the `run` method and pass this namespace down. The advantage is that we don't have to get the local namespace several times, which is relatively expensive since we have to go through a lot of objects and call `isinstance` on them to see whether we are interested in them. Actually, now that I am writing this: we should probably not filter out "uninteresting" objects beforehand, because that means we will not get a warning if we have a clash between the name of such an object (say, a string) and a variable. We could rather always transmit the full namespace and then raise an error when such an object is actually used as a variable name (or a warning for a name clash). Either way, I think the change is a good one because it makes the code simpler (less `level=level+1` or `level=level+2`) and wastes less time in the mildly expensive namespace resolution. The only disadvantage is that we can't make a difference in warning messages anymore between clashes that occur with names in the local namespace vs. clashes with names in the explicitly provided `run(..., namespace={...})` namespace. But then, I don't think this leads to any confusion.
2. We no longer try to analyse certain strings (thresholds, resets, refractory condition) at creation time (where this check would only succeed if the expressions do not use external variables that are unknown at that time). This has clearer disadvantages: we don't get error messages for incorrect syntax / unit errors, etc. at creation time but only at the time of before run. The advantages: The code is clearer (no `try/except` around some checks) and we don't do things like unit checking twice. I am not 100% sure about this one, but with the new error messages that point to the creation statement, I think this is ok. It is mostly in interactive use that this could be annoying. But then, with some refactoring that better separate the various stages of code generation, we could at least deal with syntax changes directly where they occur again.
3. Another big change: Instead of two methods, a `can_integrate` function that returns a `bool` and a `__call__` function that creates the actual abstract code, state updaters now only have a single `__call__` function that will raise an `UnsupportedEquationsException` if it cannot integrate them. The reason for this is that for the more complicated state updaters (e.g. linear), the only robust way of determining whether it can integrate given equations is by actually trying it (this is what we did in `LinearStateUpdater.can_integrate`). Therefore, we actually generated the code twice for examples like CUBA, once for checking whether we can use `linear` and once for actually using it. I don't really see a disadvantage of the new approach, it is pretty straightforward. We had a feature before where you could tell Brian to use a state updater despite its claiming not to be able to integrate the given equations which we lose now, but I don't think that had any practical use.

Now, the cumulative effect of these changes is not great, but it is something. If I do a `run(0*ms)` and remove the connection code (which otherwise dominate the total runtime), then the total run time for CUBA goes down from 4.2s to 3.4s on my machine, and from 3.9s to 3.1s for COBAHH. If I use an artificial example that makes use of code generation and namespace resolution alot like this:
```Python
G = NeuronGroup(100, 'v : 1')
for x in xrange(100):
    G.v['i == x'] = 'x'
```
then the reduction is more significant, it goes down from 6.5s to 2.3s (but of course here the state updater stuff does not play a role).

Thoughts?