@thesamovar 2015-11-13
This all looks pretty good to me. Some comments:

For (1) namespace clash errors, could we perhaps store alongside the namespace dict, a ``namespace_origin`` dict that has the same keys but the values are where the name came from, e.g. "local namespace", "specified in NeuronGroup call", etc. Like that, when a clash happens we can give a meaningful error message. 

(2) seems good to me, agreed that these early checks aren't so important with the new error reporting thing.

(3) is good too. In retrospect, integrating the equations twice is not very sensible! :)

I'm pretty happy for this to be merged once tests pass, but what do you think about the idea for namespace clash errors?