@mstimberg 2015-11-13
> For (1) namespace clash errors, could we perhaps store alongside the namespace dict, a namespace_origin dict that has the same keys but the values are where the name came from, e.g. "local namespace", "specified in NeuronGroup call", etc. Like that, when a clash happens we can give a meaningful error message.

This is a good point (nothing that actually changed in this branch, though), we partly do this already, but only for clashes between variables in different namespaces, not between internal variables and variables in the namespace. What changed in this branch is that we cannot give different warnings for the following two situations anymore (which we did not do before either, though...):
```Python
# Situation 1
foo = 3
G = NeuronGroup(1, 'foo:1', threshold='foo>1')
run(0*ms)
# Situation 2
G = NeuronGroup(1, 'foo:1', threshold='foo>1')
run(0*ms, {'foo': 3})
```
But as I said, I think using the term "run namespace" in both cases should not lead to confusion. I will try to improve the error message in general, though.

> (3) is good too. In retrospect, integrating the equations twice is not very sensible! :)

Well, in the beginning we did some analysis in the first phase instead, e.g. tried to check whether the equations are linear etc. But it turned out that there are several things that can go wrong along the way, things that were to complex for sympy etc. so that the only reliable way to find out whether the non-explicit state updaters can integrate given equations is to actually do it...