@mstimberg 2015-11-13
Ok, I pushed a change that should improve the error messages. For this lovely bit of code:
```Python
foo = 3
volt = 7
abs = 10
pi = 4
G = NeuronGroup(1, '''foo : 1
                      bar : 1''',
                namespace={'bar': 5, 'abs': 8},
                threshold='(foo + pi)*volt>abs(bar*volt)')
run(0*ms)
```
we now get these warnings:
```
WARNING  brian2.groups.group.Group.resolve.resolution_conflict: The name "volt" refers to different objects in different namespaces used for resolving names in the context of group "neurongroup". Will use the object from the units namespace with the value volt, but the name also refers to a variable in the run namespace with value 7.
WARNING  brian2.groups.group.Group.resolve.resolution_conflict: The name "pi" refers to different objects in different namespaces used for resolving names in the context of group "neurongroup". Will use the object from the constants namespace with the value 3.141592653589793, but the name also refers to a variable in the run namespace with value 4.
WARNING  brian2.groups.group.Group.resolve.resolution_conflict: "bar" is an internal variable of group "neurongroup", but also exists in the group-specific namespace with the value 5. The internal variable will be used.
WARNING  brian2.groups.group.Group.resolve.resolution_conflict: "foo" is an internal variable of group "neurongroup", but also exists in the run namespace with the value 3. The internal variable will be used.
WARNING  brian2.groups.group.Group.resolve.resolution_conflict: The name "abs" refers to different objects in different namespaces used for resolving names in the context of group "neurongroup". Will use the object from the functions namespace with the value <Function>, but the name also refers to a variable in the following namespaces: group-specific, run.
```
Maybe we could make the text more specific for the function, but I think it is already quite helpful like this.