@thesamovar 2015-11-13
Great, those error messages are really useful. I think this is good to merge once AppVeyor finishes.